separate cli from requirable server
